fix(trusty-console): self-heal Degraded handles + review nits (#1159)
Addresses trusty-review findings on PR #1161:

[MEDIUM] Self-healing Degraded state: `ensure_connected` now treats
`HandleState::Degraded` like `None` once the `SpawnBackoff` window elapses,
dropping back to `None` and re-running the full `initialize` + `tools/list`
probe. While inside the backoff window the handle still returns
`McpHandleError::Degraded` (UI keeps showing the badge + hint). When marking
Degraded from the tools/list probe, `record_failure` is used instead of
`reset` so the window is non-zero (at least BACKOFF_BASE_MS = 1 s before
any re-probe). Two new tests verify the in-window and post-window paths:
`mcp_handle_degraded_within_backoff_window_stays_degraded` and
`mcp_handle_degraded_self_heals_after_backoff_window`.

[LOW] Silent no-op for missing handle keys: `run_serve` now emits
`tracing::warn!` in the `None` branches for `trusty-memory` and
`trusty-search` handle lookups so a future key rename is immediately visible.

[LOW] UI fallback hint: changed the Svelte nullish-fallback string in
`ServiceCard.svelte` from a near-copy of the Rust `DEGRADED_HINT` to a
generic "Service reachable but its metrics tool is unavailable." to avoid
a two-place sync hazard. UI dist rebuilt via pnpm build.

Serde: `ServiceStatus` already carries `#[serde(rename_all = "snake_case")]`;
added a new `service_status_serialises_to_lowercase_strings` test in
`connector.rs` to lock the wire contract defensively.

File split: `mcp_handle.rs` was 556 SLOC (> 500 prod cap). Converted to
`mcp_handle/mod.rs` + `mcp_handle/tests.rs` (basename `tests.rs` qualifies
for the 1500-SLOC test budget per #1131). No allowlist entries changed.
